Changed Layer class __repr__ method usage
The "num_input" attribute is now no longer printed by default since it is largely irrelevant for neural networks. This caused the InputLayer class no longer having to override this method.

Usage of the `repr()` function to get the `name` instance attribute value, and the representation for next layer of the network were both removed. `!r` is now used to ensure calling __repr__, not __str__.