Fix big mistake during DenseLayer back propagation
In the same for-loop where the gradients of the layer's output to the layer's input is calculated, the network was already updates. This means that for every data instance the layer was updates, essentially creating a different model. Causing the new gradients to not be for the same layer of the model anymore.

Also extracted the `scaled_alpha * gradient[o]` calculation outside the for-loop, causing a lot less calculations/computations to be performed.